# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Primeiro Prontuário API
  version: 1.0.0
  description: |
    Contrato completo do MVP acadêmico para um único médico.
    Use somente dados fictícios ou anonimizados.
  license:
    name: Não licenciado - uso acadêmico demonstrativo
    identifier: LicenseRef-Proprietary
servers:
  - url: https://prontuario.local
    description: Cluster Kubernetes local com TLS confiado apenas na máquina de demonstração
...
security:
  - sessionCookie: []
paths:
...
  /api/v1/patients/{patientId}/consultations:
    parameters:
      - $ref: "#/components/parameters/PatientId"
    post:
      tags: [Consultas]
      operationId: createConsultationDraft
      summary: Criar rascunho de consulta
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationDraftRequest"
      responses:
        "201":
          description: Rascunho criado
          headers:
            Location:
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/consultations/{consultationId}:
    parameters:
      - $ref: "#/components/parameters/ConsultationId"
    get:
      tags: [Consultas]
      operationId: getConsultation
      summary: Consultar rascunho ou consulta finalizada
      responses:
        "200":
          description: Consulta, incluindo adendos quando existentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
    put:
      tags: [Consultas]
      operationId: updateConsultationDraft
      summary: Atualizar rascunho clínico
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConsultationRequest"
      responses:
        "200":
          description: Rascunho atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/consultations/{consultationId}/finalization:
    parameters:
      - $ref: "#/components/parameters/ConsultationId"
    post:
      tags: [Consultas]
      operationId: finalizeConsultation
      summary: Finalizar consulta de forma imutável
      security:
        - sessionCookie: []
          csrfHeader: []
      responses:
        "200":
          description: Consulta finalizada; repetição idempotente indica alreadyFinalized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/consultations/{consultationId}/addenda:
    parameters:
      - $ref: "#/components/parameters/ConsultationId"
    post:
      tags: [Consultas]
      operationId: addConsultationAddendum
      summary: Adicionar adendo append-only
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddendumRequest"
      responses:
        "201":
          description: Adendo criado
          headers:
            Location:
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Addendum"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/patients/{patientId}/medical-record:
    parameters:
      - $ref: "#/components/parameters/PatientId"
    get:
      tags: [Prontuário]
      operationId: getMedicalRecord
      summary: Ler prontuário finalizado em ordem cronológica
      parameters:
        - name: from
          in: query
          schema:
            type: string
            examples: ["2030-01-15T10:00:00"]
          description: Data local inclusiva, sem offset
        - name: to
          in: query
          schema:
            type: string
            examples: ["2030-01-15T11:00:00"]
          description: Data local exclusiva, sem offset
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Página cronológica de consultas finalizadas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecordPage"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Sessão server-side; HttpOnly, SameSite=Lax e Secure em HTTPS
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: |
        Token devolvido por GET /api/v1/auth/csrf e associado ao cookie
        XSRF-TOKEN do mesmo cliente. Exigido em mutações autenticadas.
  headers:
    CorrelationId:
      description: UUID de correlação aceito ou gerado pela API
      schema:
        type: string
        format: uuid
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    PatientId:
      name: patientId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ConsultationId:
      name: consultationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Problem:
      description: Problem Details seguro; status e tipo dependem da falha
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    ConsultationStatus:
      type: string
      enum: [DRAFT, FINALIZED]
    ConsultationDraftRequest:
      type: object
      additionalProperties: false
      properties:
        appointmentId:
          type: [string, "null"]
          format: uuid
        anamnesis:
          type: [string, "null"]
        chiefComplaint:
          type: [string, "null"]
        physicalExamination:
          type: [string, "null"]
        diagnosticHypotheses:
          type: [string, "null"]
        treatmentPlan:
          type: [string, "null"]
        observations:
          type: [string, "null"]
        clinicalDate:
          type: [string, "null"]
          description: Data local sem offset
          examples: ["2030-01-15T10:00:00"]
    UpdateConsultationRequest:
      type: object
      required: [version]
      additionalProperties: false
      properties:
        anamnesis:
          type: [string, "null"]
        chiefComplaint:
          type: [string, "null"]
        physicalExamination:
          type: [string, "null"]
        diagnosticHypotheses:
          type: [string, "null"]
        treatmentPlan:
          type: [string, "null"]
        observations:
          type: [string, "null"]
        version:
          type: integer
          format: int64
          minimum: 0
    AddendumRequest:
      type: object
      required: [content, justification]
      additionalProperties: false
      properties:
        content:
          type: string
          minLength: 1
        justification:
          type: string
          minLength: 1
    Addendum:
      type: object
      required: [id, consultationId, content, justification, authorId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        consultationId:
          type: string
          format: uuid
        content:
          type: string
        justification:
          type: string
        authorId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    Consultation:
      type: object
      required:
        - id
        - patientId
        - status
        - version
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        appointmentId:
          type: [string, "null"]
          format: uuid
        anamnesis:
          type: [string, "null"]
        chiefComplaint:
          type: [string, "null"]
        physicalExamination:
          type: [string, "null"]
        diagnosticHypotheses:
          type: [string, "null"]
        treatmentPlan:
          type: [string, "null"]
        observations:
          type: [string, "null"]
        status:
          $ref: "#/components/schemas/ConsultationStatus"
        finalizedBy:
          type: [string, "null"]
          format: uuid
        finalizedAt:
          type: [string, "null"]
          format: date-time
        version:
          type: integer
          format: int64
        alreadyFinalized:
          type: [boolean, "null"]
        addenda:
          type: [array, "null"]
          items:
            $ref: "#/components/schemas/Addendum"
    MedicalRecordAddendum:
      type: object
      required: [id, content, justification, authorId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        justification:
          type: string
        authorId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    MedicalRecordItem:
      type: object
      required:
        - id
        - clinicalDate
        - createdAt
        - finalizedBy
        - finalizedAt
        - addenda
      properties:
        id:
          type: string
          format: uuid
        clinicalDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        anamnesis:
          type: [string, "null"]
        chiefComplaint:
          type: [string, "null"]
        physicalExamination:
          type: [string, "null"]
        diagnosticHypotheses:
          type: [string, "null"]
        treatmentPlan:
          type: [string, "null"]
        observations:
          type: [string, "null"]
        finalizedBy:
          type: string
          format: uuid
        finalizedAt:
          type: string
          format: date-time
        addenda:
          type: array
          items:
            $ref: "#/components/schemas/MedicalRecordAddendum"
    MedicalRecordPage:
      allOf:
        - $ref: "#/components/schemas/PageMetadata"
        - type: object
          required: [content]
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/MedicalRecordItem"
...
    PageMetadata:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    Problem:
      type: object
      required: [type, title, status, detail, correlationId]
      properties:
        type:
          type: string
          format: uri
          examples: [urn:problem:invalid-request]
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        correlationId:
          type: string
          format: uuid
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"